metiq: add handling of broken frame at a matching point
For now search backwards at most half a frame duration for a valid value
This means that the maximum error is half the frame duration.
However, if the limit would be hit the overall quality of the measurement
is probably so bad that the result is questionable already.
Add debug flag to verify.py so if something goes wrong it is visible.

Signed-off-by: JohanBlome <jblome@fb.com>