metiq: move audio offset compensation This place is more obvious and also allow for calculations even when the delay values are close to the offset
Signed-off-by: JohanBlome <jblome@fb.com>

metiq: move audio offset compensation
This place is more obvious and also allow for calculations even
when the delay values are close to the offset

To test this particular situaion the unittest definitions and
generation code had to be changed in a manner to allow for a
more flexible test definition.
Test 12 is the test for this.

Signed-off-by: JohanBlome <jblome@fb.com>