metiq: refactor lock-layout
Previously lock-layout locked on the first succesful tag identification,
and kept that state in vft.

This is now simplified using a two pass process where the video is first
searched for a valid tag setup. This is then used for the entire clip.
A benefit of this approach is that more frames before the good tag setup
can be parsed than before.

Tag detection can be flaky if the image is not in a good condition
(sharp, high contrast etc) where the gray code parsing is robust.
The only potential problem is if the first detction is a bad.
In that case running without lock-layout may be needed (or a different
mechanism for finding a better candidate).

Signed-off-by: JohanBlome <jblome@fb.com>