metiq: refactored lock-layout
Prviously lock-layout locked on the first succesful tag identification
and kept that state in vft.
This is now simplified to a two pass process where the video is first
searched for a valid tag setup and that is then used for the entire clip.
A benefit of tis is that more frames before the good tag setup can be parsed
than using the old mechanism.

Tag detection can be flakey if the image is not in a good condition
(sharp, high contrast etc) where the gray code parsing is robust.
The only potential problem is if the first detction is a bad.
In that case running without lock-layout may be needed (or a different
mechanism for finding a better candidate).

Signed-off-by: JohanBlome <jblome@fb.com>