metiq: add manual tagging possibility
video _analyze.py has a new cli arg:
"--tag-manual" which opens up a window where coordinates can
be set using mouse.

This way files that are a little blurry and/or challenging in
another way can still be parsed.

The coords need to be set in order (as of now) with
1) top left
2) top right
3) bottom left
4) bottom right

Signed-off-by: JohanBlome <jblome@fb.com>