metiq: add parse error filtering to video analysis
If running at double the display frequency it is impossible to have
a single value between other adjacent values. Remove thise obvious
errors.

In order to have valid data the capturing frequency shoudl always be
at least twice the display frequency. If that is not the case we should
not filter.

Signed-off-by: JohanBlome <jblome@fb.com>