metiq: add cli options to enable bandpass
Off should be default.

The filter order was too high and the low frequency threshold also too high.
Adjusted these to be much more gentle values to not causing problems in good
conditions.

Better to not filter unless needed.

There are a few ways to combat low audio correlation and few hits.
1) move the microphone and speakers so they are pointing towards eachother.
   Otherwise directivy can cause filtered capture.
2) Enable the bandpass filter for noiose captures.
3) Lower threshold for identified source. This can introduce false positives.

Signed-off-by: JohanBlome <jblome@fb.com>