metiq: fix issue in the scipy audio reader
The scipy audio reader is producing negative audio sample numbers:

```
$ ./src/metiq.py parse -i test.mp4
$ csvlook -I test.mp4.audio.csv
| audio_sample | timestamp  | correlation        |
| ------------ | ---------- | ------------------ |
| -2013.0      | -0.1258125 | 33.661738326336064 |
| 45236.0      | 2.82725    | 43.675673512366174 |
| 92211.0      | 5.7631875  | 43.36033385786903  |
| 140211.0     | 8.7631875  | 44.65195202003493  |
| 188211.0     | 11.7631875 | 44.978145190449496 |
| 236211.0     | 14.7631875 | 42.31786780469706  |
```

The issue is caused by the get_correlation_indices function:

```
  # add a needle length to the front
  haystack = np.concatenate((np.zeros(needlesize), haystack))
```

This pads the haystack with zeros at the beginning so that correlations
can be found at the very start of the audio. Later we return the results
as follows:
```
  return [
      [peak - needlesize, cc]
      for peak, cc in zip(peaks, corrcoeff)
      if cc > min_match_threshold
  ]
```

The code subtracts `needlesize` to compensate for the padding. However, if
a correlation peak is found within the padded zeros region (i.e.
peak < needlesize), the result will be negative.

This happens when there is a partial match at the very beginning of the
audio: The correlation finds a peak in the zero-padded region, which after
subtracting needlesize produces a negative sample number.

In our case, the -2013.0 sample at timestamp -0.1258125 seconds indicates
the algorithm detected a partial match that would theoretically have started
before the audio file actually begins. This is a false positive caused by
either:

1. Noise at the start of the recording matching partially with the needle
2. The correlation detecting what it thinks is a chirp signal in the
padding/early audio

Note that this one may be correct: Note that the distance between the negative
timestamp and the first positive timestamp is ~3 seconds. Still, we prefer not
to have them.

The fix consists of filtering out negative indices before returning results,
since they represent impossible positions (before the audio started):
```
  return [
      [peak - needlesize, cc]
      for peak, cc in zip(peaks, corrcoeff)
      if cc > min_match_threshold and peak >= needlesize
  ]
```

Tested:
```
$ ./src/metiq.py parse -i test.mp4
$ csvlook -I test.mp4.audio.csv
| audio_sample | timestamp  | correlation        |
| ------------ | ---------- | ------------------ |
| 45236.0      | 2.82725    | 43.675673512366174 |
| 92211.0      | 5.7631875  | 43.36033385786903  |
| 140211.0     | 8.7631875  | 44.65195202003493  |
| 188211.0     | 11.7631875 | 44.978145190449496 |
| 236211.0     | 14.7631875 | 42.31786780469706  |
```